+++ NEWS.md
@@ -1,4 +1,5 @@
 # lintr 1.0.3.9000 # 
+* New exports: ids_with_token() and with_id() (#297 @stufield)
 * Add support for overriding GitHub API Token via `GITHUB_TOKEN` environment
   variable (#63, @mattyb)
 * Changed the default value of the `length` argument to `object_length_linter` to 30 for consistency (#325 @DragosMG)